+++ src/vsr/versor_motor_pybind11.cpp
@@ -35,7 +35,7 @@ void AddMotor(py::module &m) {
       .def_buffer([](Mot &arg) -> py::buffer_info {
         return py::buffer_info(arg.data(), sizeof(double),
                                py::format_descriptor<double>::value(), 1,
-                               {arg.Num}, {sizeof(double)});
+                               {static_cast<unsigned long>(arg.Num)}, {sizeof(double)});
       });
 }
