+++ src/vsr/versor_rotor_pybind11.cpp
@@ -34,7 +34,7 @@ void AddRotor(py::module &m) {
       .def_buffer([](Rot &arg) -> py::buffer_info {
         return py::buffer_info(arg.data(), sizeof(double),
                                py::format_descriptor<double>::value(), 1,
-                               {arg.Num}, {sizeof(double)});
+                               {static_cast<unsigned long>(arg.Num)}, {sizeof(double)});
       });
 }
