+++ src/vsr/versor_tangent_vector_pybind11.cpp
@@ -44,7 +44,7 @@ void AddTangentVector(py::module &m) {
       .def_buffer([](Tnv &arg) -> py::buffer_info {
         return py::buffer_info(arg.data(), sizeof(double),
                                py::format_descriptor<double>::value(), 1,
-                               {arg.Num}, {sizeof(double)});
+                               {static_cast<unsigned long>(arg.Num)}, {sizeof(double)});
       });
 }
