+++ src/vsr/versor_vector_pybind11.cpp
@@ -40,7 +40,7 @@ void AddVector(py::module &m) {
       .def_buffer([](Vec &arg) -> py::buffer_info {
         return py::buffer_info(arg.data(), sizeof(double),
                                py::format_descriptor<double>::value(), 1,
-                               {arg.Num}, {sizeof(double)});
+                               {static_cast<unsigned long>(arg.Num)}, {sizeof(double)});
       });
 }
